chore: function names cleanup
